Added Vibrato and updated bool size in msg protocol

+++ ESP32_Client_V1.0/src/Extras/LocalStorage.h
@@ -30,8 +30,8 @@ public:
     void SetDeviceName(std::string name);
     uint16_t GetDeviceID();
     void SetDeviceID(uint16_t id);
-    uint8_t GetDeviceBoolean();
-    void SetDeviceBoolean(uint8_t deviceBoolean);
+    uint16_t GetDeviceBoolean();
+    void SetDeviceBoolean(uint16_t deviceBoolean);
     uint8_t GetNumOfDistributors();
     void SetNumOfDistributors(uint8_t numOfDistributors);
     void GetDistributorConstruct(uint16_t distributorNum, uint8_t* construct);
